+++ RELEASE_NOTES.md
@@ -3,3 +3,5 @@ v0.67.6
 - Retire old SVR2 staging enclave
 
 - keytrans: Remove unused APIs
+
+- backups: Relax check on session switchover update message authors